Split atmosphereTasks into separate files
